Create MPPT_P&O
Written a program in embedded C language using the MPLAB X Integrated Development Environment (IDE) to enable integration with Proteus software. The program is designed to simulate the behaviour of a DC motor that is powered by a Solar Panel. This simulation can help in testing the functionality of the DC motor and its interaction with the Solar Panel under different conditions. By running the program in Proteus, one can obtain valuable insights into how the DC motor would behave in real-world scenarios.